+++ src/lighthouse.php
@@ -49,8 +49,8 @@ return [
     |--------------------------------------------------------------------------
     |
     | The guard to use for authenticating GraphQL requests, if needed.
-    | This setting is used whenever Lighthouse looks for an authenticated
-    | user, for example in directives such as @guard or @can.
+    | This setting is used whenever Lighthouse looks for an authenticated user, for example in directives
+    | such as `@guard` and when applying the `AttempAuthentication` middleware.
     | TODO this setting will default to 'api' in v5
     |
     */