+++ tests/Integration/IntrospectionTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Integration;
 
+use Nuwave\Lighthouse\Schema\RootType;
 use Nuwave\Lighthouse\Schema\TypeRegistry;
 use Tests\TestCase;
 use Tests\Utils\Scalars\Email;
@@ -27,7 +28,7 @@ class IntrospectionTest extends TestCase
 
     public function testFindsTypesFromSchema(): void
     {
-        $this->schema .= '
+        $this->schema .= /** @lang GraphQL */ '
         type Foo {
             bar: Int
         }
@@ -37,7 +38,7 @@ class IntrospectionTest extends TestCase
             $this->introspectType('Foo')
         );
         $this->assertNotNull(
-            $this->introspectType('Query')
+            $this->introspectType(RootType::QUERY)
         );
 
         $this->assertNull(