+++ tests/Utils/Models/Category.php
@@ -12,6 +12,5 @@ use Illuminate\Database\Eloquent\Model;
  */
 class Category extends Model
 {
-    protected $guarded = [];
     protected $primaryKey = 'category_id';
 }