+++ tests/Utils/Models/Image.php
@@ -17,8 +17,6 @@ use Illuminate\Database\Eloquent\Relations\MorphTo;
  */
 class Image extends Model
 {
-    protected $guarded = [];
-
     public function imageable(): MorphTo
     {
         return $this->morphTo();