+++ src/Execution/Arguments/ArgPartitioner.php
@@ -9,6 +9,7 @@ use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\Relations\MorphOne;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
+use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Support\Contracts\ArgResolver;
 use Nuwave\Lighthouse\Support\Utils;
 use ReflectionClass;
@@ -185,6 +186,10 @@ class ArgPartitioner
 
         if (! $returnType instanceof ReflectionNamedType) {
             return false;
+        }
+
+        if (! class_exists($returnType->getName())) {
+            throw new DefinitionException('Class '.$returnType->getName().' does not exist, did you forget to import the Eloquent relation class?');
         }
 
         return is_a($returnType->getName(), $relationClass, true);