+++ tests/Utils/Models/Role.php
@@ -15,8 +15,6 @@ class Role extends Model
 {
     public $timestamps = false;
 
-    protected $guarded = [];
-
     public function users(): BelongsToMany
     {
         return $this