+++ CHANGELOG.md
@@ -9,6 +9,29 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- Allow passing additional headers to `multipartGraphQL` test helper https://github.com/nuwave/lighthouse/pull/1342
+- Add empty root types automatically when extending them https://github.com/nuwave/lighthouse/pull/1347
+- Configure a default `guard` for all authentication functionality https://github.com/nuwave/lighthouse/pull/1343
+
+### Changed
+
+- Improve validation error when extending a type that is not defined https://github.com/nuwave/lighthouse/pull/1347
+- Use `forceFill()` when mutating models https://github.com/nuwave/lighthouse/pull/1348
+
+## 4.12.4
+
+### Fixed
+
+- Fix nesting OR within AND condition when using `@whereConditions` https://github.com/nuwave/lighthouse/pull/1341
+
+## 4.12.3
+
+### Changed
+
+- Throw an exception if the return type declaration class for a relation does not exist https://github.com/nuwave/lighthouse/pull/1338
+
 ## 4.12.2
 
 ### Fixed